Added debug tool & EventManager singleton method changed
C-A-M-l : object info copy
C-A-M-r : object remove

EventManager instance is saved into a global-accessible pointer.
Mouse event now includes button number.

A bug-like behaviour found, that main scene animation is stopped even
when popup message is called. need to be fixed later.

+++ src/Graphic.cpp
@@ -417,7 +417,7 @@ void Graphic::DrawQuad(const VertexInfo *vi) { DrawQuads(vi, 4); }
 // This method should be called every time to calculate FPS
 void Graphic::BeginFrame()
 {
-  frame_delay_weight_avg_ = frame_delay_weight_avg_ * 0.8f + delta() * 1000.0f * 0.2f;
+  frame_delay_weight_avg_ = frame_delay_weight_avg_ * 0.8f + (float)delta() * 1000.0f * 0.2f;
   last_render_time_ = Timer::GetUncachedSystemTime();
 
   // remove rare gap of render time (about 20FPS)
